requirements.txt
psycopg2-binary
pandas
requests
python-dotenv
sqlalchemy
lightgbm
scikit-learn
numpy
joblib
fastapi
uvicorn
pydantic
streamlit
plotly
psycopg2-binary
